+++ RELEASE_NOTES.md
@@ -25,6 +25,7 @@
     * Renamed `Procrastinator::Config#provide_context` to `provide_container`
     * You must now call `Scheduler#work` on the result of `Procrastinator.config`
     * Use a dedicated process monitor (like `monit`) instead in production environments
+    * Suuply a block to `daemonized!` to run code in the spawned process.
 * `max_tasks` is removed as it only added concurrency complexity
 
 ### Minor Changes